refactored code into smaller files
