+++ src/bigcommerce_plugin.js
@@ -3,8 +3,12 @@ import optimizedOnePageCheckout from './layout_config/optimized_one_page_checkou
 import onePageCheckout from './layout_config/one_page_checkout'
 import addressBook from './layout_config/address_book'
 import createAccount from './layout_config/create_account'
+import IdentifyLayouts from "./identify_layouts";
 
-export default class BigCommercePlugin {
+window.AF = window.AF || {}
+window.AF.BigCommercePlugin = BigCommercePlugin
+
+class BigCommercePlugin {
   constructor(widgetConfig){
     this.version = "1.3.0"
     this.widgetConfig = widgetConfig
@@ -14,6 +18,34 @@ export default class BigCommercePlugin {
       ...addressBook,
       ...createAccount
     ]
-    new MutationHelper(this.layoutConfigurations, this.widgetConfig)
+    new IdentifyLayouts(this.layoutConfigurations, this.widgetConfig)
   }
 }
+
+// this is configured using:
+//
+// window.AddressFinderConfig = {
+//   key: "ADDRESSFINDER_DEMO_KEY",
+//   widgetOptions: {
+//     byline: false
+//   },
+//   debug: true
+// }
+
+
+let _initPlugin = function(){
+  window.AF._plugin = new BigCommercePlugin({
+    nzKey: window.AddressFinderConfig.key_nz || window.AddressFinderConfig.key || window.AddressFinderConfig.key_au,
+    auKey: window.AddressFinderConfig.key_au || window.AddressFinderConfig.key || window.AddressFinderConfig.key_nz,
+    nzWidgetOptions: window.AddressFinderConfig.nzWidgetOptions || window.AddressFinderConfig.widgetOptions || {},
+    auWidgetOptions: window.AddressFinderConfig.auWidgetOptions || window.AddressFinderConfig.widgetOptions || {},
+    debug: window.AddressFinderConfig.debug || false
+  })
+}
+
+let s = document.createElement('script')
+s.src = 'https://api.addressfinder.io/assets/v3/widget.js'
+s.async = 1
+s.onload = _initPlugin
+document.body.appendChild(s)
+
