+++ src/mutation_helper.js
@@ -6,70 +6,24 @@ import "core-js/fn/array/from"
 import "core-js/fn/array/includes"
 import "core-js/fn/string/includes"
 
-import FormHelper from "./form_helper"
-
 export default class MutationsHelper {
-  constructor(layoutConfigurations, widgetConfig) {
-    this.formHelperConfig = {}
-    this.formHelpers = []
+  constructor({layoutConfigurations, widgetConfig, formHelperConfig, formHelpers, initialiseFormHelper}) {
+    this.formHelpers = formHelpers
+    this.formHelperConfig = formHelperConfig
     this.layoutConfigurations = layoutConfigurations
     this.widgetConfig = widgetConfig
+    this.initialiseFormHelper = initialiseFormHelper
     this.countryCodes = ["au", "nz"]
-    this.identifyLayout()
     this.monitorMutations()
   }
 
-  identifyLayout(){
-    for (const layoutConfig of this.layoutConfigurations) {
-      let identifyingElement = document.querySelector(layoutConfig.layoutSelector)
+  identifyAdditionalLayouts(){
+    const layoutSelectorExists = config => document.querySelector(config.layoutSelector)
+    const isNewFormHelper = config => !this.anyFormHelpersWithLayoutIdentifier(config.layoutSelector)
 
-      if (identifyingElement) {
-        this._log(`Identified layout named: ${layoutConfig.label}`)
-        this.initialiseFormHelper(layoutConfig)
-      }
-    }
-  }
-
-  initialiseFormHelper(layoutConfig){
-    let searchElement = document.getElementById(layoutConfig.searchIdentifier)
-
-    if (searchElement) {
-      this.formHelperConfig = {
-        countryElement: document.getElementById(layoutConfig.countryIdentifier),
-        label: layoutConfig.label,
-        layoutSelector: layoutConfig.layoutSelector,
-        nz: {
-          countryValue: layoutConfig.nz.countryValue,
-          searchElement: document.getElementById(layoutConfig.nz.elements.address1),
-          elements: {
-            address_line_1_and_2: document.getElementById(layoutConfig.nz.elements.address1),
-            address_line_1: null,
-            address_line_2: null,
-            suburb: document.getElementById(layoutConfig.nz.elements.suburb),
-            city: document.getElementById(layoutConfig.nz.elements.city),
-            region: document.getElementById(layoutConfig.nz.elements.region),
-            postcode: document.getElementById(layoutConfig.nz.elements.postcode)
-          },
-          regionMappings: null
-        },
-        au: {
-          countryValue: layoutConfig.au.countryValue,
-          searchElement: document.getElementById(layoutConfig.au.elements.address1),
-          elements: {
-            address_line_1_and_2: null,
-            address_line_1: document.getElementById(layoutConfig.au.elements.address1),
-            address_line_2: document.getElementById(layoutConfig.au.elements.address2),
-            locality_name: document.getElementById(layoutConfig.au.elements.suburb),
-            city: null,
-            state_territory: document.getElementById(layoutConfig.au.elements.state),
-            postcode: document.getElementById(layoutConfig.au.elements.postcode)
-          },
-          stateMappings: layoutConfig.au.stateMappings
-        }
-      }
-      let helper = new FormHelper(this.widgetConfig, this.formHelperConfig)
-      this.formHelpers.push(helper)
-    }
+    this.layoutConfigurations.filter(layoutSelectorExists)
+                             .filter(isNewFormHelper)
+                             .forEach(this.initialiseFormHelper.bind(this))
   }
 
     // search active formHelpers for this layoutSelector
@@ -138,17 +92,8 @@ export default class MutationsHelper {
     return true
   }
 
-  identifyAdditionalLayouts(){
-    const layoutSelectorExists = config => document.querySelector(config.layoutSelector)
-    const isNewFormHelper = config => !this.anyFormHelpersWithLayoutIdentifier(config.layoutSelector)
-
-    this.layoutConfigurations.filter(layoutSelectorExists)
-                             .filter(isNewFormHelper)
-                             .forEach(this.initialiseFormHelper.bind(this))
-  }
-
   _formHelpersWithMissingElements(){
-    const hasMissingElements = formHelper => !this.areAllElementsStillInTheDOM()
+    const hasMissingElements = () => !this.areAllElementsStillInTheDOM()
 
     return this.formHelpers.filter(hasMissingElements)
   }
@@ -166,11 +111,73 @@ export default class MutationsHelper {
     this.identifyAdditionalLayouts()
   }
 
-  _log(message, object1=undefined){
+  _mutationHandler(mutations){
+    const changedNodes = mutations.reduce((nodes, mutation) => {
+      // ignore this mutation if the target is the AddressFinder UL element
+      if (mutation.target && mutation.target.classList && mutation.target.classList.contains("af_list")) {
+        return nodes
+      }
+
+      return nodes.concat([...mutation.addedNodes]).concat([...mutation.removedNodes])
+    }, [])
+
+    const anyBigCommerceChanges = changedNodes.find((node) => {
+      return !(node.classList && node.classList.contains("af_list"))
+    })
+
+    if (!anyBigCommerceChanges) {
+      return // ignore AddressFinder changes
+    }
+
+    if (this._mutationTimeout) {
+      clearTimeout(this._mutationTimeout) // reset previous timeout
+    }
+
+    // ignore any further changes for the next 750 mS
+    this._mutationTimeout = setTimeout(this.resetAndReloadFormHelpers.bind(this), 750)
+
+    // this._setMutationTimeout()
+  }
+
+  _domNodeModifiedHandler(event){
+    if (event.target.className && event.target.className.includes("af_list")) {
+      return // ignore AddressFinder changes
+    }
+
+    if (event.relatedNode && event.relatedNode.className && event.relatedNode.className.includes("af_list")) {
+      return // ignore AddressFinder changes
+    }
+
+    if (this._mutationTimeout) {
+      clearTimeout(this._mutationTimeout) // reset previous timeout
+    }
+
+    // ignore any further changes for the next 750 mS
+    this._mutationTimeout = setTimeout(this.resetAndReloadFormHelpers.bind(this), 750)
+  }
+
+  monitorMutations(){
+    if (window.MutationObserver) {
+      /* for modern browsers */
+      var observer = new MutationObserver(this._mutationHandler.bind(this));
+      observer.observe(document.body, {childList: true, subtree: true});
+
+    } else if (window.addEventListener) {
+      /* for IE 9 and 10 */
+      document.body.addEventListener('DOMNodeInserted', this._domNodeModifiedHandler.bind(this), false);
+      document.body.addEventListener('DOMNodeRemoved', this._domNodeModifiedHandler.bind(this), false);
+    } else {
+        if (window.console) {
+          console.info('AddressFinder Error - please use a more modern browser')
+        }
+    }
+  }
+
+  _log(message, data=undefined){
     // widgetConfig.debug should be on 
-    if (window.console) {
-      if (object1 != undefined) {
-        console.log(`${message}`, object1)
+    if (this.widgetConfig.debug && window.console) {
+      if (data != undefined) {
+        console.log(`${message}`, data)
       }
       else {
         console.log(`${message}`)
